requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
distro==1.9.0
exceptiongroup==1.2.1
groq==0.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
lxml==5.2.2
numpy==2.0.0
pandas==2.2.2
pydantic==2.7.4
pydantic_core==2.18.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sniffio==1.3.1
typing_extensions==4.12.2
tzdata==2024.1
